Started Cleaning up caption package
