--- app/src/main/java/com/storyspots/caption/Screen.kt
@@ -1,6 +1,0 @@
-data class ScreenBounds(
-    val north: Double,
-    val south: Double,
-    val east: Double,
-    val west: Double
-)
